+++ packages/stargate/src/signingstargateclient.spec.ts
@@ -49,6 +49,15 @@ describe("SigningStargateClient", () => {
           ],
           gas: "160000",
         },
+        transfer: {
+          amount: [
+            {
+              amount: "4000",
+              denom: "ucosm",
+            },
+          ],
+          gas: "160000",
+        },
         undelegate: {
           amount: [
             {
@@ -107,6 +116,15 @@ describe("SigningStargateClient", () => {
           ],
           gas: "160000",
         },
+        transfer: {
+          amount: [
+            {
+              amount: "502400",
+              denom: "utest",
+            },
+          ],
+          gas: "160000",
+        },
         undelegate: {
           amount: [
             {
@@ -156,6 +174,15 @@ describe("SigningStargateClient", () => {
             },
           ],
           gas: "120000",
+        },
+        transfer: {
+          amount: [
+            {
+              amount: "4000",
+              denom: "ucosm",
+            },
+          ],
+          gas: "160000",
         },
         undelegate: {
           amount: [
@@ -207,6 +234,15 @@ describe("SigningStargateClient", () => {
           ],
           gas: "160000",
         },
+        transfer: {
+          amount: [
+            {
+              amount: "502400",
+              denom: "utest",
+            },
+          ],
+          gas: "160000",
+        },
         undelegate: {
           amount: [
             {
@@ -253,7 +289,6 @@ describe("SigningStargateClient", () => {
 
       // got tokens
       const after = await client.getBalance(beneficiaryAddress, "ucosm");
-      assert(after);
       expect(after).toEqual(transferAmount[0]);
     });
 
@@ -280,7 +315,6 @@ describe("SigningStargateClient", () => {
 
       // got tokens
       const after = await client.getBalance(beneficiaryAddress, "ucosm");
-      assert(after);
       expect(after).toEqual(transferAmount[0]);
     });
   });
@@ -465,8 +499,6 @@ describe("SigningStargateClient", () => {
                 assert(customDelegatorAddress, "missing customDelegatorAddress");
                 assert(customValidatorAddress, "missing validatorAddress");
                 assert(customAmount, "missing amount");
-                assert(customAmount.amount, "missing amount.amount");
-                assert(customAmount.denom, "missing amount.denom");
                 return {
                   delegator_address: customDelegatorAddress,
                   validator_address: customValidatorAddress,
@@ -733,8 +765,6 @@ describe("SigningStargateClient", () => {
                 assert(customDelegatorAddress, "missing customDelegatorAddress");
                 assert(customValidatorAddress, "missing validatorAddress");
                 assert(customAmount, "missing amount");
-                assert(customAmount.amount, "missing amount.amount");
-                assert(customAmount.denom, "missing amount.denom");
                 return {
                   delegator_address: customDelegatorAddress,
                   validator_address: customValidatorAddress,